pruebas y fix de errores en usuario y examenes

+++ FransfordSystem/Models/Descripcion.cs
@@ -27,11 +27,11 @@ namespace FransfordSystem.Models
 
         [Display(Name = "Valor mínimo")]
         [RegularExpression(@"^-?(([1-9]\d*)|0)(.0*[1-9](0*[1-9])*)?$", ErrorMessage = "Valor invalido, debe ser de 0 a 15")]
-        public int? valorMinimo { get; set; }
+        public float? valorMinimo { get; set; }
 
         [Display(Name = "Valor máximo")]
         [RegularExpression(@"^-?(([1-9]\d*)|0)(.0*[1-9](0*[1-9])*)?$", ErrorMessage = "Valor invalido, debe ser de 0 a 15")]
-        public int? valorMaximo { get; set; }
+        public float? valorMaximo { get; set; }
 
         [Display(Name = "Unidades")]
         [Required(ErrorMessage = "Este campo es obligatorio")]
